- Added an api for race car data

+++ SignalR/Models/Speedometer.cs
@@ -8,7 +8,7 @@ namespace DAS_Server_SignalR.SignalR.Models
         public string RaceId { get; set; } = default!;
 
         [JsonPropertyName("speed")]
-        public int Speed { get; set; }
+        public decimal Speed { get; set; }
 
         [JsonPropertyName("raceTime")]
         public long RaceTime { get; set; }
